Renamed `secrets.py` to `secret.py`
`secrets` is already a Python standard library module. Hence, it should be
renamed to something less ambiguous. Using  `secret.py`. The other files that use
or mention this file have been updated. `isort` was also run on the
individual files since they were previously sorted assuming that `secrets` was
in reference to the standard library.