Refactor `spotify.py` to only add new tracks
Before, on each iteration of the script, we were removing and replacing all tracks in the user's script unconditionally. Now, we check the tracks in the user's playlist against the tracks returned by the Last.fm API to only add tracks that weren't already in the playlist. Thus, history is preserved and only new tracks are added.